Add mutex to mempool
