+++ domain/miningmanager/mempool/mempool_utxo_set.go
@@ -24,6 +24,7 @@ func newMempoolUTXOSet(mp *mempool) *mempoolUTXOSet {
 	}
 }
 
+// this function MUST be called with the mempool mutex locked for writes
 func (mpus *mempoolUTXOSet) addTransaction(transaction *model.MempoolTransaction) {
 	outpoint := &externalapi.DomainOutpoint{TransactionID: *transaction.TransactionID()}
 
@@ -42,6 +43,7 @@ func (mpus *mempoolUTXOSet) addTransaction(transaction *model.MempoolTransaction
 	}
 }
 
+// this function MUST be called with the mempool mutex locked for writes
 func (mpus *mempoolUTXOSet) removeTransaction(transaction *model.MempoolTransaction) {
 	for _, input := range transaction.Transaction().Inputs {
 		mpus.poolUnspentOutputs[input.PreviousOutpoint] = input.UTXOEntry
@@ -56,6 +58,7 @@ func (mpus *mempoolUTXOSet) removeTransaction(transaction *model.MempoolTransact
 	}
 }
 
+// this function MUST be called with the mempool mutex locked for reads
 func (mpus *mempoolUTXOSet) checkDoubleSpends(transaction *externalapi.DomainTransaction) error {
 	outpoint := externalapi.DomainOutpoint{TransactionID: *consensushashing.TransactionID(transaction)}
